feat: change website purpose
Change the repository's name to "site-wolfdaemon" in order to
facilitate the repurposing of the site. The site is now a
homepage for all of wolfdaemon's activities, with the
original site being repurposed into a page within the site.

Add a consistent navbar for both pages. Make it so that both
pages have a styling that is consistent to the original site.